Moving the module suite project to a subdirectory, as it should have been years ago (having projects inside projects is a no-no in NetBeans). Had to move the branding directory as well.
Changed core and cmd modules to be plain Java projects, with no NetBeans
connection at all. Moved the parser module pieces to the core, which was
the only module using it. Removed the remains of the parser module.

Changed nbcore module to be a psuedo library wrapper for the core, cmd,
and parser jars. Everything that depended on core/cmd/parser now depends
on the nbcore module instead.

Note that branding is not working entirely, will fix later.


git-svn-id: file:///Users/nfiedler/jswat.svn/trunk@107 6539ee5b-5550-41f0-90e5-622d21111933