Avoid NPE in NetBeans project support code
Apparently something changed in the latest version of NetBeans such
that the projects code expects to find real projects. Since that is
a great deal of work to implement, avoid invoking the API in NetBeans
that results in the NPE, namely GlobalPathRegistry. This seems to have
very little effect, at least for simple Java code.