Use NetBeans installer instead of our own
Remove everything that has to do with our customer installer and
build system. No longer need the custom launchers, the script for
producing the list of required modules (and hence the list of
disabled modules), nor the top-level build script.

Hard coding the version number in the properties file.

Updated all of the documentation.

[issue #5]