Fix classpath registration, no more errors
Restored the registration of the classpath and sourcepath via the
GlobalPathRegistry in NetBeans. The errors around the classpath and
sourcepath are resolved with the addition of a ClassPathProvider
implementation. Open projects are provided via the trampoline (though
only one is ever returned). However, none of this seems to help the
Java editor syntax highlighting, which is broken even in the
Preferences dialog.